refactor: move index inside manifest
