+++ commons/src/main/java/io/aiven/kafka/tieredstorage/commons/manifest/SegmentManifestV1.java
@@ -19,7 +19,7 @@ package io.aiven.kafka.tieredstorage.commons.manifest;
 import java.util.Objects;
 import java.util.Optional;
 
-import io.aiven.kafka.tieredstorage.commons.index.ChunkIndex;
+import io.aiven.kafka.tieredstorage.commons.manifest.index.ChunkIndex;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonInclude;
