+++ commons/src/main/java/io/aiven/kafka/tieredstorage/commons/manifest/index/ChunkIndex.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.tieredstorage.commons.index;
+package io.aiven.kafka.tieredstorage.commons.manifest.index;
 
 import java.util.List;
