+++ commons/src/main/java/io/aiven/kafka/tieredstorage/commons/manifest/index/FixedSizeChunkIndexBuilder.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.tieredstorage.commons.index;
+package io.aiven.kafka.tieredstorage.commons.manifest.index;
 
 public class FixedSizeChunkIndexBuilder extends AbstractChunkIndexBuilder {
     private final int transformedChunkSize;
