+++ commons/src/test/java/io/aiven/kafka/tieredstorage/commons/manifest/index/VariableSizeChunkIndexBuilderTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.tieredstorage.commons.index;
+package io.aiven.kafka.tieredstorage.commons.manifest.index;
 
 import io.aiven.kafka.tieredstorage.commons.Chunk;
