+++ commons/src/test/java/io/aiven/kafka/tieredstorage/commons/manifest/index/serde/ChunkSizesBinaryCodecTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.aiven.kafka.tieredstorage.commons.index.serde;
+package io.aiven.kafka.tieredstorage.commons.manifest.index.serde;
 
 import java.nio.ByteBuffer;
 import java.util.ArrayList;
